[BE Admin - Product] - Tự Hiếu, bắt lỗi form, chưa hiển thị thông báo xác nhận

+++ laptech/src/main/java/com/laptech/rest/admin/ProductRestController.java
@@ -5,6 +5,7 @@ import java.util.Iterator;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -73,9 +74,32 @@ public class ProductRestController {
     @DeleteMapping("{id}")
     public ResponseEntity<Void> delete(@PathVariable("id") Long id){
         if(!dao.existsById(id)){
-             return ResponseEntity.notFound().build();
+            return ResponseEntity.notFound().build();
+        } else {
+            Product prod = dao.findById(id).get();
+            if(!prod.getOrderDetails().isEmpty()) {
+                 return new ResponseEntity<>(HttpStatus.CONFLICT);
+            } else {
+                List<ProductImages> listPImg = pImgDao.findByProduct(prod);
+                List<Price> listPrice = priceDao.findByProduct(prod);
+                if(!listPImg.isEmpty()) {
+                    for (ProductImages pimg : listPImg) {
+                    pImgDao.deleteById(pimg.getId());
+                    fileService.delete("images", pimg.getName());
+                }
+                }
+                if(!listPrice.isEmpty()) {
+                  for (Price p : listPrice) {
+                    priceDao.deleteById(p.getId());
+                   }
+                }
+               
+                
+                
+                dao.deleteById(id);
+            }
         }
-        dao.deleteById(id);
+       
         return ResponseEntity.ok().build();
     }
